requirements.txt
fastapi-slim==0.111.0
httpx==0.27.0
pydantic-settings==2.3.0
python-dotenv==1.0.1
